Added paste event for pasting images while writing the blog

+++ services/crud/src/main/java/com/blog/crud/blog/BlogService.java
@@ -6,7 +6,10 @@ import com.blog.crud.constants.Constants;
 import com.blog.crud.exceptions.BlogNotFoundException;
 import com.blog.crud.exceptions.CategoryNotFoundException;
 import com.blog.crud.exceptions.UnauthorizedException;
+import com.blog.crud.payments.PaymentClient;
 import com.blog.crud.utils.Encryption;
+
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
@@ -18,15 +21,25 @@ public class BlogService {
     private final CategoryRepository categoryRepository;
     private final BlogRepository blogRepository;
     private final Encryption encryptor;
+    private final PaymentClient paymentClient;
 
-    public BlogResponse getBlog(String encodedId) {
+    public BlogResponse getBlog(String encodedId, String username) {
+        // Add logic to check if the blog that is being fetched is under subscription or not 
         var id = encryptor.decodeId(encodedId);
         var blog = blogRepository.findById(id)
                 .orElseThrow(() -> new BlogNotFoundException(Constants.BLOG_DOES_NOT_EXISTS));
+        
+        var isSubscribed = paymentClient.getPaymentStatus(username); 
+        if (isSubscribed) {
+            return blogMapper.toBlogResponse(blog);
+        }
+        blog.setContent(blog.getContent().substring(100));
+
         return blogMapper.toBlogResponse(blog);
     }
 
     public String createBlog(BlogRequest blogRequest) {
+        // Add logic to only allow users with premium subscription to create more than 10 blogs
         var category = categoryRepository.findById(Integer.valueOf(blogRequest.categoryId()))
                 .orElseThrow(() -> new CategoryNotFoundException(Constants.CATEGORY_DOES_NOT_EXISTS));
         var blog = blogRepository.save(blogMapper.toBlog(blogRequest, category));
@@ -38,17 +51,18 @@ public class BlogService {
                 .orElseThrow(() -> new CategoryNotFoundException(Constants.CATEGORY_DOES_NOT_EXISTS));
         var blog = blogRepository.findById(encryptor.decodeId(blogRequest.id()))
                 .orElseThrow(() -> new BlogNotFoundException(Constants.BLOG_DOES_NOT_EXISTS));
-        //! Check if the user that has written the blog is updating the same 
         validateOwnership(blog, username);
         mergeBlog(blog, blogRequest);
         blogRepository.save(blog);
     }
 
-    public void deleteBlog(String encodedId) {
+    public void deleteBlog(String encodedId, String username) {
         var id = encryptor.decodeId(encodedId);
         var blog = blogRepository.findById(id)
                 .orElseThrow(() -> new BlogNotFoundException(Constants.BLOG_DOES_NOT_EXISTS));
-        blogRepository.delete(blog);
+        if (blog.getAuthor().equals(username)) {
+            blogRepository.delete(blog);
+        }
     }
 
     private void validateOwnership(Blog blog, String username) {
